refactor: remove unnecessary code
