+++ app/components/sections/AddEditSection.tsx
@@ -1,4 +1,4 @@
-import { Form, useActionData, useTransition } from '@remix-run/react'
+import { Form, useActionData, useSubmit, useTransition } from '@remix-run/react'
 import { useEffect, useState } from 'react'
 import Button from '../form/Button'
 import { trimValue } from '~/utils'
@@ -6,7 +6,6 @@ import { useTranslation } from 'react-i18next'
 import DialogWrapper from '../Dialog'
 import type { ActionData } from '~/routes/sections'
 export interface editItem {
-  id: string
   name: string
   description: string
 }
@@ -14,22 +13,37 @@ const AddEditSection = ({
   open,
   setOpen,
   editItem,
-  addSection,
-  editSection,
+  editId,
 }: {
   open: boolean
   setOpen: (e: boolean) => void
   showErrorMessage?: boolean
   editItem?: editItem
+  editId?: string
   addSection?: (name: string, description: string) => void
-  editSection?: (name: string, description: string) => void
 }) => {
   const { t } = useTranslation()
   const transition = useTransition()
   const [sectionName, setSectionName] = useState('')
   const [description, setDescription] = useState('')
+  const submit = useSubmit()
   const sectionActionData = useActionData() as ActionData
-
+  const editSection = (name: string, description: string) => {
+    submit(
+      {
+        editSection: 'sectionEdit',
+        id: editId!,
+        name: name,
+        description: description,
+      },
+      {
+        method: 'post',
+      }
+    )
+  }
+  const addSection = (name: string, description: string) => {
+    submit({ addSection: 'sectionAdd', name, description }, { method: 'post' })
+  }
   useEffect(() => {
     if (editItem) {
       setSectionName(editItem.name)
@@ -49,7 +63,7 @@ const AddEditSection = ({
     }
   }, [sectionActionData?.resp, setOpen, t])
   const handleEdit = (name: string, description: string) => {
-    editSection?.(name, description)
+    editSection(name, description)
   }
   const handleAdd = (name: string, description: string) => {
     addSection?.(name, description)
